Started moving over to postcode instead of stopcode
The program now takes and validates a postcode as user input, rather than a stop code. Next steps are to get lon and lat from the postcode and store them within the Location struct. This will then used to get arrivals at the two nearest stoppoints to the given postcode